+++ user-service/lib/query/users/index.js
@@ -1,9 +1,11 @@
 const registraton = require('./registraton');
 const search = require('./search');
 const authentication = require('./authentication');
+const update = require('./update');
 
 module.exports = {
   registraton,
   search,
   authentication,
+  update,
 };