+++ user-service/lib/query/users/registraton.js
@@ -5,23 +5,35 @@ const { sendAuthEmail } = require('../../mailer/mail-defintion');
 const isEmail = require('validator/lib/isEmail');
 const isURL = require('validator/lib/isURL');
 const isNumeric = require('validator/lib/isNumeric');
+const isDate = require('validator/lib/isDate');
 
-function isFormValid(
+function isCompanyFormValid(
   email,
   letterOK,
   companyNumber,
   companyPhone,
   companyHomepage,
+  companyLogo,
 ) {
   return (
     isEmail(email) &&
-    typeof letterOK === 'boolean' &&
+    isDate(letterOK, { locale: 'ko-KR' }) &&
     companyNumber.length === 10 &&
     isNumeric(companyPhone) &&
     (companyPhone.length === 11 ||
       companyPhone.length === 10 ||
       companyPhone.length === 8) &&
-    isURL(companyHomepage)
+    (!companyHomepage || isURL(companyHomepage)) &&
+    (!companyLogo || isURL(companyLogo))
+  );
+}
+
+function isSocialFormValid(email, profile_image, thumbnail_image, letter_ok) {
+  return (
+    isEmail(email) &&
+    isURL(profile_image) &&
+    isURL(thumbnail_image) &&
+    isDate(letter_ok, { locale: 'ko-KR' })
   );
 }
 
@@ -30,7 +42,7 @@ function isFormValid(
  * @typedef {{
  * email: string
  * plain_password: string
- * letter_ok: boolean
+ * letter_ok: Date
  * }} RegistratonInformation
  */
 
@@ -43,18 +55,67 @@ function isFormValid(
  * company_address: string
  * company_owner: string
  * company_homepage: string
+ * company_logo: string
  * }} CompanyInformation
  */
 
 /**
  * 소셜 가입 정보
  * @typedef {{
+ * social_id: string
+ * token: string
  * email: string
  * nickname: string
  * profile_image: string
  * thumbnail_image: string
+ * letter_ok: Date
  * }} SocialInformation
  */
+
+/**
+ *
+ * @param {SocialInformation} socialInfo
+ * @returns {Promise<boolean>} 회원가입 성공 여부
+ */
+async function registerSocial(socialInfo) {
+  try {
+    // form validation
+    if (
+      !isSocialFormValid(
+        socialInfo.email,
+        socialInfo.profile_image,
+        socialInfo.thumbnail_image,
+        socialInfo.letter_ok,
+      )
+    ) {
+      return false;
+    }
+    const responses = await query([
+      {
+        sql: 'INSERT INTO social_profile VALUES (?, ?, ?, ?, ?, ?, ?, ?)',
+        values: [
+          socialInfo.social_id,
+          socialInfo.token,
+          socialInfo.email,
+          socialInfo.nickname,
+          socialInfo.profile_image,
+          socialInfo.thumbnail_image
+            ? socialInfo.thumbnail_image
+            : socialInfo.profile_image,
+          new Date(),
+          socialInfo.letter_ok ? socialInfo.letter_ok : null,
+        ],
+      },
+    ]);
+    if (responses[0].rows.affectedRows > 0) {
+      return true;
+    } else {
+      return false;
+    }
+  } catch (e) {
+    throw e;
+  }
+}
 
 /**
  *
@@ -65,23 +126,15 @@ function isFormValid(
  */
 async function registerCompany(regInfo, companyInfo) {
   try {
-    // 회사 정보
-    const {
-      company_number,
-      company_name,
-      company_phone,
-      company_address,
-      company_owner,
-      company_homepage,
-    } = companyInfo;
     // 입력된 폼이 유효한지 검사합니다.
     if (
-      isFormValid(
+      isCompanyFormValid(
         regInfo.email,
         regInfo.letter_ok,
-        company_number,
-        company_phone,
-        company_homepage,
+        companyInfo.company_number,
+        companyInfo.company_phone,
+        companyInfo.company_homepage,
+        companyInfo.company_logo,
       )
     ) {
       const saltRounds = 10;
@@ -91,7 +144,7 @@ async function registerCompany(regInfo, companyInfo) {
       const token = randomstring.generate(30);
 
       // 회원 가입
-      await query([
+      const responses = await query([
         // 업주 등록 쿼리
         {
           sql: 'INSERT INTO registration VALUES (?, ?, ?, ?, ?)',
@@ -100,20 +153,21 @@ async function registerCompany(regInfo, companyInfo) {
             hash_string,
             null,
             'company',
-            regInfo.letter_ok ? new Date() : null,
+            regInfo.letter_ok ? regInfo.letter_ok : null,
           ],
         },
         // 업주 프로필 생성 쿼리
         {
-          sql: 'INSERT INTO company_profile VALUES (?, ?, ?, ?, ?, ?, ?)',
+          sql: 'INSERT INTO company_profile VALUES (?, ?, ?, ?, ?, ?, ?, ?)',
           values: [
             regInfo.email,
-            company_number,
-            company_name,
-            company_phone,
-            company_address,
-            company_owner,
-            company_homepage,
+            companyInfo.company_number,
+            companyInfo.company_name,
+            companyInfo.company_phone,
+            companyInfo.company_address,
+            companyInfo.company_owner,
+            companyInfo.company_homepage ? companyInfo.company_homepage : null,
+            companyInfo.company_logo ? companyInfo.company_logo : null,
           ],
         },
         // 이메일 인증 토큰 생성 쿼리
@@ -122,13 +176,17 @@ async function registerCompany(regInfo, companyInfo) {
           values: [regInfo.email, token, 'reg', null], // 이메일 인증 토큰은 만료기간이 없습니다.
         },
       ]);
-      // SMTP로 인증 메일을 보냅니다.
-      await sendAuthEmail(
-        regInfo.email,
-        `"Digicu" <${process.env.SMTP_FROM}>`,
-        token,
-      );
-      return true;
+      if (responses[0].rows.affectedRows > 0) {
+        // SMTP로 인증 메일을 보냅니다.
+        await sendAuthEmail(
+          regInfo.email,
+          `"Digicu" <${process.env.SMTP_FROM}>`,
+          token,
+        );
+        return true;
+      } else {
+        return false;
+      }
     } else {
       return false;
     }
@@ -137,10 +195,9 @@ async function registerCompany(regInfo, companyInfo) {
   }
 }
 
-async function registerUser(regInfo, socialInfo) {}
-
 async function registerAdmin(regInfo) {}
 
 module.exports = {
   registerCompany,
+  registerSocial,
 };